Update ember-cli version.
