+++ bower.json
@@ -5,24 +5,24 @@
     "aptible-sass": "aptible/aptible-sass#diesel",
     "bootstrap": "~3.3.2",
     "bootstrap-sass": "twbs/bootstrap-sass#3.3.1",
-    "ember": "1.13.5",
+    "ember": "1.13.6",
     "ember-cli-shims": "ember-cli/ember-cli-shims#0.0.3",
     "ember-cli-test-loader": "ember-cli-test-loader#0.1.3",
     "ember-data": "1.0.0-beta.18",
-    "ember-resolver": "~0.1.15",
+    "ember-resolver": "~0.1.18",
     "ember-load-initializers": "ember-cli/ember-load-initializers#0.1.4",
-    "ember-qunit": "0.4.6",
+    "ember-qunit": "0.4.9",
     "ember-qunit-notifications": "0.0.7",
     "font-awesome": "~4.3.0",
     "jquery": "^1.11.1",
     "loader.js": "ember-cli/loader.js#3.2.0",
     "nouislider": "~7.0.9",
     "pretender": "~0.6.0",
-    "qunit": "~1.17.1",
+    "qunit": "~1.18.0",
     "zeroclipboard": "~2.2.0",
     "JavaScript-MD5": "~1.1.0"
   },
   "resolutions": {
-    "ember": "1.13.5"
+    "ember": "1.13.6"
   }
 }
