+++ package.json
@@ -23,8 +23,8 @@
     "broccoli-asset-rev": "^2.0.2",
     "broccoli-sass": "^0.6.6",
     "core-object": "^1.1.0",
-    "ember-cli": "0.2.7",
-    "ember-cli-app-version": "0.3.3",
+    "ember-cli": "1.13.6",
+    "ember-cli-app-version": "0.4.0",
     "ember-cli-aptible-shared": "0.0.26",
     "ember-cli-babel": "^5.0.0",
     "ember-cli-content-security-policy": "0.4.0",
@@ -35,21 +35,22 @@
     "ember-cli-flash": "^1.3.3",
     "ember-cli-font-awesome": "0.0.4",
     "ember-cli-gravatar": "^1.2.1",
-    "ember-cli-htmlbars": "0.7.6",
+    "ember-cli-htmlbars": "0.7.9",
     "ember-cli-htmlbars-inline-precompile": "^0.2.0",
-    "ember-cli-ic-ajax": "0.1.1",
+    "ember-cli-ic-ajax": "0.2.1",
     "ember-cli-inject-live-reload": "^1.3.0",
     "ember-cli-pretender": "^0.3.1",
-    "ember-cli-qunit": "0.3.13",
+    "ember-cli-qunit": "0.3.20",
     "ember-cli-release": "^0.2.5",
     "ember-cli-sentry": "damiencaselli/ember-cli-sentry#2bcea789ebab6a93700c6edcb67282e0960cfad7",
+    "ember-cli-sri": "^1.0.1",
     "ember-cli-template-lint": "0.1.1",
     "ember-cli-uglify": "^1.0.1",
     "ember-data": "1.0.0-beta.18",
     "ember-data-hal-9000": "0.1.7",
     "ember-deploy-s3": "0.0.4",
     "ember-disable-proxy-controllers": "^1.0.0",
-    "ember-export-application-global": "^1.0.2",
+    "ember-export-application-global": "^1.0.3",
     "ember-feature-flags": "1.0.0",
     "ember-radio-button": "^0.1.3",
     "ember-truth-helpers": "0.0.5",
